tui/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sawa-tui"
version = "0.2.0"
description = "Bloomberg-style TUI for browsing S&P 500 data"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "sawa",
    "rich>=13.0",
    "blessed>=1.20",
    "psycopg[binary]>=3.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "pillow>=10.0",
    "cairosvg>=2.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "mypy>=1.0",
    "ruff>=0.1.0",
]

[project.scripts]
sawa-tui = "sawa_tui.app:main"

[tool.hatch.build.targets.wheel]
packages = ["sawa_tui"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
